+++ apps/web/src/pages/RegisterPage.tsx
@@ -8,7 +8,6 @@ import { useFeatureStore } from '../stores/features'
 import { isTauri, setSecureToken } from '../secureStorage'
 import { openExternalUrl } from '../openExternalUrl'
 import { ROUTES } from '../routes'
-import { requestPushNotificationPermissionIfNeeded } from '../pushNotifications'
 import { setPersistedSocialView } from '../socialView'
 import { resolvePostAuthRoute } from '../authRedirect'
 
@@ -90,7 +89,6 @@ export default function RegisterPage() {
 
         setLoading(true)
         try {
-            void requestPushNotificationPermissionIfNeeded()
             const res = await authApi.register(username, email, password, captchaToken || undefined)
             setAuth(res.token, res.user)
             setActiveDmChannelId(null)